Modify Cronofy.php
I would like to propose to use protected instead of private so that the Cronofy objects can be inherited and it's functionality can be used by the child class. I need this to store all access token data (including expires_at) in a access token object to store this in the database next to my application's user account data. But in general using protected makes it more flexible to use.